Factor home configurations
This is the beginning of a large refactor motivated by a few factors.
The first is, I'm better at nix now, and understand how modules work.
The second is, I have a Mac now, and want to be able to share
configurations without having a combinatorial explosion of conditionals
and components.
The third is, someone suggested that I reorient the configurations
around "Setting up the machine to do a thing" instead of "What is set up
on a machine that does the thing". Subtle change of perspective! I am
only starting to grasp the ramifications.

This step just cleans up the structure of the configurations, bumps my
dependencies, and pulverizes the Home modules into finer components that
are still oriented around "What is on the machine".