Implement (linux only) graphical configuration
Here I have come to the first major hurdle of the new structure:
conditional imports are fraught with nixos modules. I need to pass in
the hostPlatform data in such a way that it does not participate in the
fixpoint computation of the module set.

home-manager has a provision for that: extraSpecialArgs. However, this
does not seem readily available to the standalone installation sans
flakes. Since flakes are not fit for purpose, I will have to extend both
home-manager and basalt to handle passing this data in.

God willing, it is not too big of an effort!