Set up home-manager experiment
I had to disable my xmonad overrides because, bizarrely, it simply does
not compile anymore. This means that I have to disable pager hints to
taffybar, which is unfortunate. Anyway, that is something to deal with
another time.

I have pointed home-manager to a fork that passes extraSpecialArgs from
the commandline. I have not yet tested it, because I need to fork basalt
for that too. For now, I just want to make sure home-manager's normal
  functionality remains intact.