Pass in a build environment thru extraSpecialArgs
The experiment is a success. I have validated a way for basalt to pass
extraSpecialArgs. It is a terrible hack, so something better designed
will be necessary, but now I can pass `environment.nix` to the
home-manager configurations, and provide access to stuff that I do not
want to configure within home-manager as such. Currently, those are the
sources of my dependencies, and the build-time pkgs.

Basically, I would like to turn the configuration as-such into a dumb
instantiation of home-manager modules, and put all of my non-trivial
infrastructure configuration in environment.nix.